requirements.txt
flask==2.0.1
werkzeug==2.0.1
pandas==1.5.3
beautifulsoup4==4.9.3
requests==2.28.2
plotly==5.13.0
dash==2.9.3
python-dotenv==1.0.0
google-api-python-client==2.86.0
wordcloud==1.9.2
numpy==1.24.2
